+++ legatium-webclient-logging/src/test/kotlin/eu/inqudium/legatium/webclient/logging/IntegrationFixture.kt
@@ -1,5 +1,6 @@
 package eu.inqudium.legatium.webclient.logging
 
+import eu.inqudium.legatium.common.PeerServer
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.extension.ExtendWith